// evio_Comms_Microservice/src/utils/enums/A8000MeasurementsEnum.ts
// Attention - changing the key names because they are the names of the variables we receive from this controller in MQTT, 
// so if you change some name make sure the controller also send it that way for all devices types
export enum A8000_MEASUREMENT_ENUMS {
    Gr_ID = 'GROUP_ID',
    Op_State = 'STATE_NAME',
    A_Limit = 'CURRENT_LIMIT',
    P = 'POWER_ACTIVE',
    I_Tot = 'TOTAL_CURRENT',
    Vavg = 'VOLTAGE',
    Energy = 'ENERGY',
    Priority = 'PRIORITY',
    A_Min_Solar = 'MIN_SOLAR_CURRENT',
    Sharing_Mode = 'SHARING_MODE',
    Charging_Mode = 'CHARGING_MODE',
    Op_P_Margin = 'OPERATIONAL_MARGIN',
    EVs_Total_Psetp = 'CURRENT_DISTRIBUTED_EACH_VEHICLE',
    CB_A_Limit = 'SW_CURRENT_LIMIT',
    ILA = 'CURRENT_L1',
    ILB = 'CURRENT_L2',
    ILC = 'CURRENT_L3',
    Com_Fail = 'ERROR_COMMUNICATION',
    Circuit_Breaker = 'CIRCUIT_BREAKER',
    Active_Sessions = 'ACTIVE_SESSIONS',
    General_Alarm = 'General_Alarm',
    Com_Fault = 'Com_Fault',
    Last_Reading = 'LAST_READING',
    B_Consumption = 'LOCATION_CONSUMPTION',
    N_Phases = 'NUMBER_OF_PHASES',
    Factor = 'PV_CONVERSION_FACTOR',
}